Units: central Units module as single home for unit labels + energy toggle
Add src/core/units.{h,cpp} (namespace Units): documents SupraFit's unit
conventions (SI internally; energy display toggle; ITC mmol/L + import-time cal→J)
and centralises them. The energy-unit selection (EnergyUnit / currentEnergy()) is
relocated here from Thermo, and canonical on-screen labels (temperature,
concentration, mmol/L, µL, mL, ppm, concentration axis) get single-source helpers.
thermo/statistic format via Units::currentEnergy(); the titration concentration
charts use Units::concentrationAxis(). Behaviour-neutral (NMR/titration tests green).
