+++ AIChangelog.md
@@ -85,3 +85,4 @@ One line per significant AI-assisted improvement (newest first).
 - 2026-07-14: ITC thermodynamics — units & sign clarity. `Thermo::CurrentEnergyUnit()` (new) drives runtime-selectable output units (SI kJ/mol·J·mol⁻¹·K⁻¹ vs calorie-based kcal/mol·cal·mol⁻¹·K⁻¹) across `FormatThermo`, `MonteCarlo2Thermo`, `GridSearch2Thermo`; toggle via ConfigDialog ("Report thermodynamics in kcal/mol") → qApp `energy_unit_kcal` (persisted). Added a −TΔS row (= ΔG − ΔH) everywhere ΔH/ΔS are shown, matching the NanoAnalyze convention that caused the apparent entropy-sign mismatch.
 - 2026-07-14: Thermogram import dialog — injection volume is now editable per addition point (column 0), persisted back into `m_inject` so an `UpdateTable()` rebuild keeps manual edits (guarded by `m_updating_table`; heat columns set read-only); the cal→J scaling in `ThermogramWidget` is relabelled "cal → J" with a tooltip explaining that leaving it at 1 for calorie data is why results match calorie-based tools.
 - 2026-07-14: Config — single source of truth. `src/ui/settingsregistry.{h,cpp}` (`SupraFitSettings`) is one table of `{key, default, kind, group, label, tooltip, range, dependsOn, …}` for all 49 application settings. `SupraFitGui::ReadSettings/WriteSettings` derive the persisted-key set and defaults from it (removed the 49-entry `m_properties` list and the ~145-line default if-chain); `ConfigDialog` is generated from it (one tab per group, generic Bool/Int/Double/String widgets, `dependsOn` enable-chains, `note` section labels, `Custom` kind for the hand-built directory/working-dir section). Behaviour-neutral (same keys/defaults/ranges/dependencies); net −377 lines, configdialog.cpp 555→~230. Adding a setting is now one table row instead of four coordinated edits.
+- 2026-07-14: Units — central home. `src/core/units.{h,cpp}` (`namespace Units`) documents SupraFit's unit conventions and centralises them: the energy-unit selection (`EnergyUnit`/`currentEnergy()`, relocated from `Thermo`) and canonical on-screen label helpers (temperature/concentration/mmol·L⁻¹/µL/mL/ppm/concentration-axis). `thermo`/`statistic` now format via `Units::currentEnergy()`; the titration concentration charts use `Units::concentrationAxis()`. mmol/L for ITC syringe/cell concentrations is intentional (documented, not "fixed"); the observed-heat cal→J choice stays at import (ThermogramWidget), per the note in units.h.
